+++ src/GxIO/GxIO_MEGA_P8_MEGASHIELD_H/GxIO_MEGA_P8_MEGASHIELD_H.h
@@ -39,7 +39,7 @@ class GxIO_MEGA_P8_MEGASHIELD_H : public GxIO
     void selectRegister(bool rs_low); // for generalized readData & writeData (RA8875)
     void setBackLight(bool lit);
   private:
-    int8_t _cs, _rs, _rst, _wr, _rd, _bl; // Control lines
+    int16_t _cs, _rs, _rst, _wr, _rd, _bl; // Control lines
 };
 
 #define GxIO_Class GxIO_MEGA_P8_MEGASHIELD_H
