+++ src/GxIO/STM32MICRO/GxIO_STM32F103BluePill_P16/GxIO_STM32F103BluePill_P16.h
@@ -47,7 +47,7 @@ class GxIO_STM32F103BluePill_P16 : public GxIO
   private:
     void setDataPinsOutput();
     void setDataPinsInput();
-    int8_t _cs, _rs, _rst, _wr, _rd, _bl; // Control lines
+    int16_t _cs, _rs, _rst, _wr, _rd, _bl; // Control lines
 };
 
 #define GxIO_Class GxIO_STM32F103BluePill_P16
