+++ src/GxIO/STM32MICRO/GxIO_STM32F407ZGM4_P16/GxIO_STM32F407ZGM4_P16.h
@@ -46,7 +46,7 @@ class GxIO_STM32F407ZGM4_P16 : public GxIO
     void selectRegister(bool rs_low); // for generalized readData & writeData (RA8875)
     void setBackLight(bool lit);
   private:
-    int8_t _cs, _rs, _rst, _wr, _rd, _bl; // Control lines
+    int16_t _cs, _rs, _rst, _wr, _rd, _bl; // Control lines
     bool _bl_active_high;
 };
